[dm-migrations] Fixes CREATE TABLE for MySQL.
The CREATE TABLE statement was being output with the table options
before the column list; this was incorrect and caused a MySQL error.  As
mentioned in the Lighthouse ticket, it would be a good idea to add true
integration tests (which actually run migrations against a database) for
all supported adapters.

Credits go to Greg Campbell for the meat of this patch

[#827 state:resolved]